pyproject.toml
[tool.poetry]
name = "realt-zamok"
version = "0.1.0"
description = ""
authors = ["Mikhail Klimovich <mikhail12klimovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
bs4 = "^0.0.1"
sentry-sdk = "^0.19.5"
python-dotenv = "^0.15.0"
user_agent = "^0.1.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
